:recycle: Restructure tests - Move KPI and TokenHistory tests to ManagementAPI - Split logging- and metrics tests into subfolders - Reorder tests for ExternalTask, Logging and Metrics
