+++ test/4_external_task_api/fetch_and_lock.js
