+++ test/6_logging_api/logging_api_test.js
